backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
django-allauth
dj-rest-auth
requests
cryptography
